Simplify workspace ordering and preview backfill
